quick patch to phpembed
Summary: some testing changes I made snuck into my last checkin

Reviewed By: mathewsb

Test Plan: ran the example, usage, and test programs without incident



git-svn-id: svn+ssh://tubbs/svnroot/projects/phpembed/trunk@191325 2c7ba8d8-a2f7-0310-a573-de162e16dcc7

+++ src/php_arr.cpp
@@ -156,9 +156,9 @@ void php_array::add_assoc(char *argspec, va_list ap){
           char *arg = va_arg(ap, char *);
           if(even) {
             key = arg;
-            key_len = va_arg(ap, long);
+            key_len = va_arg(ap, unsigned int);
           } else {
-            long str_len = va_arg(ap, long);
+            unsigned int str_len = va_arg(ap, unsigned int);
             add_assoc_stringl_ex(array, realKey, key_len, arg, str_len, true);
           }
         }
